extract zones from map
