+++ components/map/Map.tsx
@@ -9,7 +9,7 @@ import {
   UserTrackingMode,
 } from '@rnmapbox/maps'
 import { MapState } from '@rnmapbox/maps/lib/typescript/components/MapView'
-import { Feature, FeatureCollection, GeoJsonProperties, Geometry } from 'geojson'
+import { Feature, GeoJsonProperties, Geometry } from 'geojson'
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { View } from 'react-native'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
@@ -22,7 +22,8 @@ import { useProcessedArcgisData } from '@/modules/map/hooks/useProcessedMapData'
 import { useScreenCenter } from '@/modules/map/hooks/useScreenCenter'
 import { colors } from '@/modules/map/utils/layer-styles/colors'
 import udrStyle from '@/modules/map/utils/layer-styles/visitors'
-import udrStyle2 from '@/modules/map/utils/layer-styles/visitors2'
+
+import MapZones from './MapZones'
 
 type Props = {
   onBottomSheetContentChange?: (content: string | null) => void
@@ -92,20 +93,6 @@ const Map = ({ onBottomSheetContentChange }: Props) => {
     return false
   }, [location])
 
-  const udrDataByPrice = useMemo(
-    () => ({
-      regular: {
-        ...udrData,
-        features: udrData?.features.filter((udr) => udr.properties?.Zakladna_cena !== 2),
-      } as FeatureCollection,
-      eur2: {
-        ...udrData,
-        features: udrData?.features.filter((udr) => udr.properties?.Zakladna_cena === 2),
-      } as FeatureCollection,
-    }),
-    [udrData],
-  )
-
   return (
     <View className="flex-1">
       <MapView
@@ -144,22 +131,6 @@ const Map = ({ onBottomSheetContentChange }: Props) => {
           minDisplacement={3}
           animated
         />
-        {udrDataByPrice.regular?.features?.length > 0 && (
-          <ShapeSource id="udrSource" shape={udrDataByPrice.regular}>
-            <FillLayer
-              id="udrFill"
-              style={udrStyle.reduce((prev, current) => ({ ...prev, ...current.paint }), {})}
-            />
-          </ShapeSource>
-        )}
-        {udrDataByPrice.eur2?.features?.length > 0 && (
-          <ShapeSource id="udrSource2" shape={udrDataByPrice.eur2}>
-            <FillLayer
-              id="udrFill2"
-              style={udrStyle2.reduce((prev, current) => ({ ...prev, ...current.paint }), {})}
-            />
-          </ShapeSource>
-        )}
         {/* {zonesData && (
           <ShapeSource id="zonesSource" shape={zonesData}>
             <FillLayer
@@ -168,6 +139,7 @@ const Map = ({ onBottomSheetContentChange }: Props) => {
             />
           </ShapeSource>
         )} */}
+        {udrData && <MapZones udrData={udrData} />}
         {selectedPolygon && (
           <ShapeSource id="highlight" shape={selectedPolygon}>
             <FillLayer
